Fixed JPA

+++ JavaEE/4WebFrameworks/Play/Play1/OficialDocumentation/7_JPA_persistence/persistance/app/models/Item.java
@@ -8,5 +8,7 @@ import play.db.jpa.Model;
 public class Item  extends Model{
 	
 	public String name;
-	public int count;
+	public Long count;
+	
+	
 }
